fix: Add missing error types to VM. Remove old error and lib mods.
